fix: updating imports, renaming folders
