+++ src/lib/add_base_element_to_head.js
