+++ src/lib/comparison_data_point.js
